Feat: Added more spaghetti

+++ src/main/java/de/tinf22b6/dhbwhub/repository/LogtableRepository.java
@@ -8,6 +8,8 @@ import de.tinf22b6.dhbwhub.repository.interfaces.SpringLogtableCommentRepository
 import de.tinf22b6.dhbwhub.repository.interfaces.SpringLogtableEventCommentRepository;
 import de.tinf22b6.dhbwhub.repository.interfaces.SpringLogtableEventPostRepository;
 import de.tinf22b6.dhbwhub.repository.interfaces.SpringLogtablePostRepository;
+import jakarta.persistence.EntityExistsException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
@@ -19,6 +21,8 @@ public class LogtableRepository {
     private final SpringLogtableCommentRepository commentRepository;
     private final SpringLogtableEventPostRepository eventPostRepository;
     private final SpringLogtableEventCommentRepository eventCommentRepository;
+    private static final String ENTRY_ALREADY_LOGGED = "Entry already logged!";
+    private static final String ENTRY_DOESNT_EXIST = "Entry does not exist!";
 
     public LogtableRepository(@Autowired SpringLogtablePostRepository postRepository,
                               @Autowired SpringLogtableCommentRepository commentRepository,
@@ -76,36 +80,75 @@ public class LogtableRepository {
         return eventCommentRepository.findByEventCommentIdAndUserId(eventCommentId,userId).orElse(null);
     }
 
-    public LikeLogtablePost savePost(LikeLogtablePost post) {
-        return postRepository.save(post);
+    public void savePost(LikeLogtablePost post) {
+        if (checkIfPostExists(post.getPost().getId(), post.getUser().getId())){
+            throw new EntityExistsException(ENTRY_ALREADY_LOGGED);
+        }
+        postRepository.save(post);
     }
 
-    public LikeLogtablePostComment saveComment(LikeLogtablePostComment comment) {
-        return commentRepository.save(comment);
+    public void saveComment(LikeLogtablePostComment comment) {
+        if (checkIfCommentExists(comment.getComment().getId(), comment.getUser().getId())){
+            throw new EntityExistsException(ENTRY_ALREADY_LOGGED);
+        }
+        commentRepository.save(comment);
     }
 
-    public LikeLogtableEventPost saveEvent(LikeLogtableEventPost eventPost) {
-        return eventPostRepository.save(eventPost);
+    public void saveEvent(LikeLogtableEventPost eventPost) {
+        if (checkIfEventPostExists(eventPost.getEventPost().getId(), eventPost.getUser().getId())){
+            throw new EntityExistsException(ENTRY_ALREADY_LOGGED);
+        }
+        eventPostRepository.save(eventPost);
     }
 
-    public LikeLogtableEventComment saveEventComment(LikeLogtableEventComment eventComment) {
-        return eventCommentRepository.save(eventComment);
+    public void saveEventComment(LikeLogtableEventComment eventComment) {
+        if (checkIfEventCommentExists(eventComment.getEventComment().getId(), eventComment.getUser().getId())){
+            throw new EntityExistsException(ENTRY_ALREADY_LOGGED);
+        }
+        eventCommentRepository.save(eventComment);
     }
 
     public void deletePost(Long id) {
+        if (!postRepository.existsById(id)){
+            throw new EntityNotFoundException(ENTRY_DOESNT_EXIST);
+        }
         postRepository.deleteById(id);
     }
 
     public void deleteComment(Long id) {
+        if (!commentRepository.existsById(id)){
+            throw new EntityNotFoundException(ENTRY_DOESNT_EXIST);
+        }
         commentRepository.deleteById(id);
     }
 
     public void deleteEventPost(Long id) {
+        if (!eventPostRepository.existsById(id)){
+            throw new EntityNotFoundException(ENTRY_DOESNT_EXIST);
+        }
         eventPostRepository.deleteById(id);
     }
 
     public void deleteEventComment(Long id) {
+        if (!eventCommentRepository.existsById(id)){
+            throw new EntityNotFoundException(ENTRY_DOESNT_EXIST);
+        }
         eventCommentRepository.deleteById(id);
     }
 
+    public boolean checkIfPostExists(Long postId, Long userId){
+        return findPost(postId, userId) != null;
+    }
+
+    public boolean checkIfCommentExists(Long commentId, Long userId){
+        return findComment(commentId, userId) != null;
+    }
+
+    public boolean checkIfEventPostExists(Long eventId, Long userId){
+        return findEventPost(eventId, userId) != null;
+    }
+
+    public boolean checkIfEventCommentExists(Long eventCommentId, Long userId){
+        return findEventComment(eventCommentId, userId) != null;
+    }
 }
